Fix loading of games.
Improve sgf parsing.
Leave it to gnugo to figure out the number of moves.
This went wrong even with the examples we ship since they contain variations.
Display more of the properties of loaded games.
Clean up the loading/starting of games and move it to the same position as the
status docker.

svn path=/trunk/KDE/kdegames/kigo/; revision=1274873