Fix the position of the initial popup message
Summary:
The initial popup message appears in the wrong position (central position
on the left side, with its left boundary outside the screen).

Add a bit of delay before displaying the message. A simple singleShot
with no delay seems to be enough to move the creation of the popup
after that the initialization of the window (personal guess) is ready.

Test Plan: The program compiles and the initial popup is shown in the correct position.

Reviewers: schwarzer, #kde_games

Reviewed By: schwarzer

Subscribers: #kde_games

Differential Revision: https://phabricator.kde.org/D5668